improve saml config options (#379)
* improve saml config options

we can set on company email domain which will be tested
when authenticating via saml and if user email matches it
it will assign the company to user.

there is also new config for custom client saml configs
which allows us to configure user specific login page which
picks the config to be used for SAML auth. then it uses
the company domain to match the config value.

STTNHUB-222